Look for Lines and MTG tester
A small app I wrote to demonstrate my ability to come up with a novel algorithm. Creates a "noisy" line of dots with zero or more bumps in it, then uses only the positions of the dots to find the bumps.

MTG tester was an app i threw together ona trip to italy. The internet there was quite bad at times, and I wanted to test out a few Magic the Gathering deck ideas. The usual online testing applications (manastack, untap, etc) were acting flaky at best, so i decided to write my own. It made for quite a fun six-hour project.